Wrap the admin using-cluster in an it block
Quick hack to pass on the port the api service is running on.
Describes "run" serially but each describe registers the function body
entirely. Thus the port is undefined at the time of attempting to use
it.

This patch fixes this issue by wrapping it in an `it` block delaying the
evaluation of the varible port.